LayoutTests/imported/w3c/web-platform-tests/fetch/sec-metadata/redirect/cross-site-redirect.tentative.https.sub-expected.txt



FAIL Cross-Site -> Same-Origin redirect assert_equals: destination expected (string) "image" but got (undefined) undefined
FAIL Cross-Site -> Same-Site redirect assert_equals: destination expected (string) "image" but got (undefined) undefined
FAIL Cross-Site -> Cross-Site redirect assert_equals: destination expected (string) "image" but got (undefined) undefined
